requirements.txt
allure-pytest==2.9.43
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
dicttoxml==1.7.4
idna==2.10
iniconfig==1.1.1
Faker==8.10.1
jsonschema==3.2.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-xdist==2.2.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.5
jsonschema==3.2.0